// src/app/models/Ereception.model.ts
import { Time } from '@angular/common';

export class EReception {
    id: number;
    Name: string;
    Gender: string;
    Mobile: string;
    Email: string;
    Address: string;
    Purpose: string;
    ToMeet: string;
    InTime: string;
    OutTime: string;

}

export class Purpose {
    value: string;
}

export class ToMeet {
    name: string;
    img: string;
}